pmda.h: refine PMDA_PMID macro
In some unrelated testing I discovered that PMDA_PMID(x,y) and
build_pmid(0,x,y) don't return the same values for some choices
of x and y!

In particular if the arguments to PMDA_PMID() are negative (!) or
larger than the corresponding field widths the PMDA_PMID() value
is clearly wrong.

Add some masking to limit x (cluster) to 12 bits and y (item) to
10 bits.