PM_ERR_BADSTORE: second round of cosmetic changes
In the places where we enumerate PM_ERR_* values and need to keep
them in sync (Perl and Python bindings in particular), make the
order the same as pmerr -l to ease maintenance.  And document this
goal in all 3 places.

Note the order in pmapi.h is different, but that serves a different
purpose and is OK to be different.