build GNUmakefiles: allow more parallel (make -j) jobs
For GNU make's -jNNN to propagate from an outer parent make to child
processes, macros that invoke $(MAKE) internally need to be decorated
with a "+" prefix.  On a 4 vcpu machine, this allows the complete pcp
tree to be built in less than half the time.

There are still some ".NOTPARALLEL:" markers in a few GNUmakefiles
that lack full dependency declarations, and subdirectory make jobs
tend to be invoked serially (due to the use of explicit shell-script
loops), so we have some further potential speedups too.