libpcp: add a pmGetOptionalConfig(3) for safe use in DSOs
There are cases both for and against the current pmGetConfig
behaviour of calling exit for terminally broken PCP installs
(missing pcp.conf, etc).  This commit resolves the issue by
adding a safe variant for third-party use, and making use of
it throughout libpcp and libpcp_pmda everywhere that shared
libraries might come a-calling.

There's a handful of libpcp call sites remaining, but these
are in specialised server or client code (logconnect and/or
pmcd, pmproxy et al) not in real need of conversion.  These
also happen to be places where new error injection is going
to be non-trivial.

Resolves Fedora BZ 1187588.