linux_proc pmda: add pmlogrewrite support for KERNEL_ULONG and jiffy-based metrics
Because the PCP type of the KERNEL_ULONG metrics is not constant
across platforms, it helps to provide pmlogrewrite rules to ensure
the type is correct, e.g. for archives that may have come from
an older platform environment (32-bit) when processed in a newer
platform environment (64-bit).

Also capture some recent metadata changes in the jiffy-based metrics.