build: add struct declarations for application headers
Add struct names for all structures, as is done in a couple of places
in pmapi.h and throughout impl.h.  This is useful to applications that
don't want to include pmapi.h and instead forward-declare structs that
they use.  The atop upstream code has headers using this model, which
we'll want to integrate with shortly.