qa: fix systemd pmda Install spurious .err output
Turned out to be a $tmp assumption, but nothing was being
done with the journalctl stderr stashed in $tmp.err anyway
so just send it to /dev/null like stdout (the return code
is all that matters here).

While there, remove some unnecessary code & tweak a couple
the delays in qa/652 to make it complete more quickly.