pcp-atop: feature-filled PMAPI version of atop
This is a new implementation of the artist formerly known as pmatop,
based on the upstream atop sources (directly).  It's thus a far more
complete version of atop, and is now a binary rather than a python
script.  Many, many additional metrics are now used by this version
compared to our original version, and these are reflected in updates
to the pmlogconf configuration.

QA work and verification of some more obscure features remain (such
as LVM/MD device handling), these will follow in subsequent commits.

Tackles Red Hat BZs 1160811, 1018575 and 1195429.