Don't fork pmcheck co-processes in a container environment.
In a container environment, pmpause is used to prevent the rc
script exiting until a SIGCHLD is received. In the rc script,
we can't run start_pmcheck in the background, otherwise when
each start_pmcheck co-process exits, pmpause would be signalled
to reap the child (start_pmcheck) rather than the grandchild
(pmlogger), and then exit - causing the whole container to exit.

	modified:   src/pmlogger/rc_pmlogger