qa/qt: add setup target for GNUmakefiles
make setup was failing ... we don't use this often (like never until
I was testing some unrelated gmake changes), but fix it to be complete.
ps. the more usual make -f GNUmakefile.install setup works because it
does not descend into the qt directory, which is another but different
issue for another commit.