pmie: prevent premature rule evaluation in pmie live mode
Resolve an issue introduced in commit d21bb6c017696 - some
code Ken and I thought was unnecessary turned out to be in
use after all.  The rest of that commit remains, it's just
the snippet of code that re-adjusts the Task scheduling vs
the current time that is affected by slow reading on stdin
(which can happen if rules are manually entered).

Added a new qa/862 test to exercise the situation and fix.