qa/260: improve stability
On slow platforms (especially VMs on low powered hosts) the pmie
startup time may be non-zero, especially where complex rules are
involved and derived metrics are in use.  This makes the time of
the first fetch variable and since the metrics are counters the
values from the firect fetch are still in use at the reporting after
the second fetch ... this makes the first results unknown (which
is deterministic) but the second results are not stable in the
presence of the rate() function because the values and timestamps
for the firtst fetch are a bit dodgey ... so filter these ones out.

After the second fetch the results are rock solid and stable
across all platforms.