Deprecate hinv.nlv and hinv.map.lvname and the LV_NAME instance domain.
The instance domain for hinv.map.lvname is the dm names e.g. dm-1,
which are not persistent. The values are the logical names, which
are persistent. These two metrics are now deprecated in favor of
hinv.map.dmname, which instead uses the persistent names for the
instance domain, and the dm names as the value of the mapping.

Note also - hinv.map.dmname is a superset of the lvm instances -
the dmname instance domain also includes non-lvm devices such as
dm-multipath, dm-cache, etc.  as well as all lvm logical devices.

The only known user of the deprecated hinv.map.lvname metric is the
old pmatop command, which itself has been deprecated.

	modified:   src/pmdas/linux/GNUmakefile
	modified:   src/pmdas/linux/clusters.h
	deleted:    src/pmdas/linux/devmapper.c
	deleted:    src/pmdas/linux/devmapper.h
	modified:   src/pmdas/linux/help
	modified:   src/pmdas/linux/indom.h
	modified:   src/pmdas/linux/pmda.c
	modified:   src/pmdas/linux/root_linux