json pmda: make build and packaging conditional
The json PMDA requires some Python modules that are not universally
available, so this commit captures that fact and makes the building
an packaging of the json PMDA conditional on the availability of
the required Python modules in the build environment.

Tested on Ubuntu 15.04 with and without the python-json-pointer
package installed and on Fedora 17 where the python-jsonpointer rpm
is not available and on RHEL 7.1 with python-jsonpointer installed.