pmnewlog.sh: fix regression
My last commit was supposed to be "cosmetic" for the usage
message, but I made a mistake and the text that was altered
was in fact a control file to pmgetopt(1) which processes
the control file to produce the usage message.

The changes caused pmgetopt(1) to fail, which cascaded into
pmnewlog failures.

This commit repairs the damage.