json pmda: tolerate json array instances that are missing -> PM_ERR_INST
Because of the caching of json array elements in the indom cache
(/var/lib/pcp/config/pmda/137.*), it is easy to get the pmda machinery
to recall indom entries that do not currently exist.  Prior to this
patch, sending those indoms to the pmda fetch-handler resulted in
needless error messages (and related slowdowns).  We now return the
correct error (PM_ERR_INST) and don't log anything.