postfix PMDA: improvements
The qshape "utility" is not always installed as an executable
on the $PATH ... so need to accommodate cases where it is a
Perl script in a backwater of the file system.

Added some logging to the pmda's log file to report which
Postfix logfile is being used and the invocation for qshape
that is being used.