pmdalinux: correct mnt/net namespace setns call issue
Resolves a problem in the networking metrics when used with
containers, reported by folks on #pcp IRC channel.  Fixes up
an issue in one section of qa/042 in the process.

Also, since mass refactoring was needed to ensure we do not
attempt to setns both mnt and net namespaces simultaneously,
I went ahead with splitting the fetch/instance refresh logic
at the same time.  Then switched over to defaulting to using
sysfs for resolution of network per-interface stats (iow, we
now prefer this over ioctl when available) - this required a
bit of additional logic to use sysfs for resolving all ioctl
metrics, which was missing previously.