New crontab entry for containers.
This crontab entry is installed by the pmlogger rc script to /etc/cron.d
on the host, which is a bind mount in the pcp-pmloger container. The host
cron service runs this to docker exec the pmlogger_daily and pmlogger_check
scripts in the running pcp-pmlogger container. If that container isn't
running, the rc script redirects all output to dev/null since there is
no reliable way of removing it on the host if the container is killed.

The rc script will not re-install the crontab entry if it already exists,
so modifications are persistent. The pcp-pmlogger Dockerfile explicitly
removes the standard /etc/cron.d/pcp-pmlogger entry after the pcp packages
are installed in the container image, so when the rc script is run for the
first time the new contab entry will be installed.

	new file:   src/pmlogger/crontab.docker.in