packaging: fix pcp-systemtools dependency on python-pcp.
pcp-systemtools depends on either or both python-pcp and python3-pcp
if they're available on the platform. There is no "one or the other"
restriction - either or both can be installed.

This fixes a reported issue where python tools in pcp-systemtools failed
to run because the platform default python2 interpreter was not installed
(even though python3 was installed).

	modified:   build/containers/pcp-collector/Dockerfile
	modified:   build/containers/pcp-monitor/Dockerfile
	modified:   build/rpm/fedora.spec
	modified:   build/rpm/pcp.spec.in