pmlogger: detect stale /var/run/pcp/pmlogger.primary.socket hard links
pmlogger.primary.socket is a hard link to the pmlogger control socket,
which has the PID of the running primary logger in it's filename.
pmlogger refuses to start if the hard link exists, even if that process
is no longer running.

This patch finds the hard link and removes it if the process is no longer
running (i.e. stale), so the primary pmlogger can still start and create
a new link. This was particularly problematic in the pcp-pmlogger container,
where such stale links abound due to the extreme prejudice that docker
uses when killing containers. If the hardlink references an actual running
primary pmlogger process, then the link is not removed and the expected
behaviour prevails. Additional QA tests for this are pending.

	modified: src/pmlogger/src/ports.c