postgresql PMDA: major overhaul
The stats views have different schemas in different versions of
postgresql ... sometimes radically so with missing columns and
added columns, and there is no attempt to add new columns at the
end of the views.

So we need a more general mapping mechanism to find the columns
we're searching for.

Also update the QA coverage in qa/778.