qa/709: revert "field overflow" changes
Revert the part of commit 68c462a that split long sequences of
numbers to address a perceived field width overflow issue.

As mamarc@amazon.com has diagnosed, this is a problem in the
base PCP Python code, and with his patch in place, the band-aid
for qa/709 is no longer required.