pmdalinux: rejig network address code for individual refresh
Refactor the pmdalinux network address metrics code along the
lines of the earlier net interface code, to ensure we only do
a minimum of refreshing, and to ensure the correct namespaces
are entered at appropriate times.

This resolves a failure in test qa/047, and is the second and
final part of the fix for SGI BZ #1067.