Fix dependency problems causing races for parallel make.
Most pmdas generate a domain.h header which contains a macro
identifying the numeric value of the domain of their metrics. In
most cases, domain.h was listed a a dependency of the pmda
itself which, in the case of parallel make, allowed it to
be generated at the same time as the source files are being
compiled, causing a race between its creation and use.

These changes make domain.h a dependency of the object files
like all the other source files and headers causing it to be
generated before compilation begins.