mingw64: access local pcp tools for cross-compiled builds
The RUN_IN_BUILD_ENV macro, which sets up the environment to
run a just-built PCP program is a problem for cross-compiling
because we cannot run the just-built tools of course.

This change allows us to use locally-installed tools (iow, of
the build-machine architecture) in the build for the tools we
need.  These are tools like newhelp, pmcpp, etc.