build: make solaris/illumos builds work without external tweaks
Extend the configure.ac script to verify the toolchain features
required to build pmmgr are present (not the case with non-gcc
compiler on Solaris).  Fix a buglet in both the pmmgr and pmwebd
checks for a c++ compiler while there.

Secondly, for Solaris we must force use of bison and flex over
the default yacc and lex, else the build fails - handle this in
the Makepkgs script.

Finally, tweak the pmdaib source for compilation out-of-the-box
on Solaris (with the default compiler).