pmwebd graphite mode: elide repeated indom searching
It was reported that for some larger archives, pmwebd was taking
noticable time running pmGetInDomArchive.  This effort was being
repeated for every metric, whereas it could be reused across metrics
with the same indom.  We now do this by memoizing the wildcard-filtered
instance names during pmns iteration.  (The cache is thrown away after
pmns traversal.)

There are no intended changes in visible behavior except for speedup;
previous QA is sufficient.