pmdalinux - add network.softnet metrics from /proc/net/softnet_stat.
/proc/net/softnet_stat exports per-CPU statistics of various interesting
metrics related to congestion in the Linux network stack. See the help
text for details. This commit adds support for these metrics summed over
all CPUs as network.softnet.*

In the future we may extend this to also export the per-cpu metrics as
network.softnet.percpu.* over the CPU instance domain - that's future work.

	modified:   src/pmdas/linux/GNUmakefile
	modified:   src/pmdas/linux/clusters.h
	modified:   src/pmdas/linux/help
	modified:   src/pmdas/linux/pmda.c
	new file:   src/pmdas/linux/proc_net_softnet.c
	new file:   src/pmdas/linux/proc_net_softnet.h
	modified:   src/pmdas/linux/root_linux