pmdate: correct year and month offset problems at end of month
Close to the end of the month (days 29, 30, 31) going forwards
or backwards by some number of months and/or years can lead to
semantic ambiguity, e.g. on 20150730 (today) what is -5 months?
20150230 is not correct, but that's what pmdate used to report.

Turns out mktime() fixes this sort of mess up, so the fix is
easy.  And the answer to the question above is 20150302.