pmdajson: default to instantaneous, no guessing at semantics
From some discussions with dsmith, default to instantaneous
semantics for all metrics - it is not possible to guess at
this using the metric type (64/double/... - unrelated).  The
pmdajson inability to distinguish semantics remains an issue
at large though, likely will need a schema extension (Davids
pondering this one - a more invasive change, so I'll leave
that for the guru).