pmlogger - also handle stale legacy hard links for primary pmlogger run files
On an upgrade from a previous version of PCP (that still used hard links for
the control and socket primary links), if pmlogger is killed with SIGKILL
prior to the upgrade, then there may be stale legacy hard links for the
primary control file and socket. If found, these need to be removed before
checking for stale symlinks - hard links are always stale here.

Also add an additional test for this in qa/808.

	modified:   qa/808
	modified:   qa/808.out
	modified:   src/pmlogger/src/ports.c