pmdamounts: add mount fs usage metrics via statvfs(3)
Extend pmdamounts to export filesystem usage metrics for the
mounts that it is tracking.  This allows one to track filesys
metrics for mounts that kernel PMDAs would typically discard,
such as for NFS clients, which in turn is useful for PCP/pmie-
based monitoring of mounts from hosts where a full PCP install
on the server is not a viable option.

Test qa/845 is extended with more coverage for pmdamounts and
to ensure the new metrics are comparable with the kernel PMDA
filesys metrics for the same mount.