__pmAddOptArchive(), __pmAdOptArchiveList(): Allow directories to be specified.
If the specified path is a directory, __pmAddOptArchive() will now
open the directory and search for archives within it. Archives are
detected by the presence of .index, .meta suffixes as well as
numeric suffixes possibly followed by one of the supported compression
suffixes.

The elements of the list provided to __pmAddOptArchiveList() may
also now be directories.

This change also reverts the previous change back to using commas
to separate the names of archives in a list. There is existing
precedent for the use of commas when PM_OPTFLAG_MULTI=true.