rc scripts: support pmlogger and pmie control file directories
This adds support for pmlogger and pmie control files to be
sourced from a directory (in addition to the single control
file model currently employed, and in a backward compatible
way).  In this way we make life easier for automation tools
like Chef, Puppet, etc when setting up monitoring farms for
remote systems.

Additionally, the work to support "sending-less-email" is now
completed here, with the addition of -l options to several of
the scripts involved, and switching to log files rather than
cron-driven emails.