rc scripts: support env var settings and local-host only mode
Add an optional server mode for critical PCP daemons (pmcd,
pmproxy, and the primary pmlogger) where we perform only the
loopback interface bind(2) instead of the allow-any model we
do by default.  This mode can be enabled using an environment
variable setting, which is propogated via the rc scripts and
files in PCP_SYSCONFIG_DIR.

Update Debian package builds so that PCP_SYSCONFIG_DIR points
to the correct location (/etc/default).  At this stage we're
not enabling this, but it appears likely Fedora might need to
run in the mode soon for the default pmcd/pmlogger processes.