pmwebd speedup: graphite mode for multiple metric/instance per archive
Reworked the inner loop of the archive fetching jobscheduler.
Formerly, each metric/instance time series was scanned individually,
(with some multithreaded parallelism).  The new code clumps together
all metrics/instances requested against the same archive, for a single
scanning loop pass.  (Multithreaded parallelism is thus exploited
across distinct archives only, which is still a win.)

This dramatically accelerates queries like the grafana default
dashboards that ask about network interfaces and memory utilization.
For example, the 3986 metrics found here for a long-span 30-day query
of network traffic for a bunch of hosts went from 50 seconds to 9.

QA is not affected, as it is a pure performance improvement.