qa: add test 816 to exercise a libpcp pdu-buffer memory issue
Exercise the libpcp pdubuf/interp issue from github issue #35.
This adds a small, anonymised archive based on the production
data sent privately in relation to the above github issue.

The test reproduces the same bug using pmdumptext, which is a
handy development as its more easily automated and it's also
alot easier to follow than the pmchart GUI code.