pcp-uptime: cater for archives without needed metrics first up
When pcp-uptime is run against an archive with an initial
result record that does not contain the metrics it wants,
but a later result record does contain 'em, some incorrect
assumptions are made and ultimately a sigsegv results.

Update test qa/742 to exercise this case alongside existing
pcp-uptime archive testing, using a new pcp-uptime2 archive.

Resolves RHBZ #1262721.