build: cater for builds with latest vector (now 1.0.2)
Switch to the more sensible model of using vector sources
directly from bintray/git instead of an intermediate tree.
In doing so, we update to 1.0.2 sources for RPM builds and
also extend the spin-rawhide script to pull latest Vector
git sources down for those Fedora rawhide builds, instead
of using a (non-vector-git, inherently out-dated) copy.

We now have the md5sum check we always should have had in
fedpkg sources.list so its easy to compare the sources we
are using for any given build to released Vector tarballs.