tsearch pdubuf: correct off-by-one error in matching address to pdubuf
The tsearch-based comparison function computed false-positives for
overlapping pdubuf address ranges for the one-past-buffer case.  We
encountered this bug on MacOS only, where memory allocations can
appear tightly packed.  This could result in prematurely freed pdubufs
and eventually worse.

We also add range-checking assertions to the result of the tsearch.