python api: ignore failures to decode instance names to utf-8
Martins came across a situation where proc metric instance names
with non-ascii characters were failing to be converted to unicode
strings in the python API.  Go with Martins simple approach of
just ignoring the failure to decode those chars, and continue on.

Adds a new test to exercise the problem scenario.  Also, audited
the rest of the module for other cases where instance names are
converted and ensured all are dealt with the same way.