libpcp: add $PCP_DEBUG to initialize pmDebug
When debugging pcp-atop it became clear the that classical -D N way
of setting PCP debug flags won't work for (an increasing set of?)
applications that have alrady laid claim to a -D flag.

So this commit provides an environment-based mechanism to initialize
pmDebug from $PCP_DEBUG if set.

Since all applications of interest use libpcp, and every application
using libpcp ends up calling __pmInitLocks() early on, and that
code has one-trip logic, then this is the right place to check the
environment and update pmDebug if needed (note update, not set pmDebug,
because the -D option processing might already have been done, so
the value from $PCP_DEBUG is bitwise OR'd in).