do not background _pmda_setup in a container environment.
In a container environment, the rc script blocks in pmpause
(as pid 1) until it is killed or all children are reaped.
If there are no children, it blocks until killed. However,
if the children fork and then exit (e.g. pmcd daemon), then
pmpause exits, killing the container with it. So we can't
background rc tasks in a container environment.