src/pmie/show.c: small changes afer FQDN changes
In showSyn() should report m->hname not m->hconn for CND_FETCH
or CND_DELAY nodes in the expression tree to match the user's
input rule (remove question in comment).

Change type of findFormat() from int to char and add the default
case for %? (action is to keep searching, which was the implied
but is now explicit).

In showSatisfyingValue() revert
    length = concat("\n<tab>   ", length, &string);
to
    length = concat("\n<space>   ", length, &string);