pmmgr: remove unhelpful comments about libpcp and containers
If ever we support nested containers, its going to be pmdaroot
that needs to handle that, not libpcp.  And in general, if you
are adding comments to code: please don't express a personal /
sarcastic opinion, and don't say "we should document X" (then
proceed to not do so).  When someone does do these things, the
comments will be out of date and need fixing too, but are quite
likely to be forgotten since they're in unrelated code.