pmlogger: change semantics for first logging operation
For each group in the configuration file (actually groups in the
configuration file with the same logging frequency are combined),
we used to add the first record to the archive after one logging
period had elapsed.

Now we add the first record for each group as soon as possible (the
log once groups get done first).

The same semantic change is implemented for dyamic changes to logging
activated via pmlc.

These changes mean we have data available earlier in the archive for
both for counter and non-counter metrics.