configure and headers: add support for strtoint64 macro
Trying to clean up the code where we call strtol() or strtoll() to
extract a number that is expected to be a 64-bit int (or
unsigned variant thereof).  The strtol() use is wrong on 32-bit
platforms.  The macros strtoint64() and strtouint64() may be
used where conversion to a 64-bit is explicit and required (e.g.
to enforce an API, or PDU, or archive format constraint).