Basic checklist code to analyze performance issues
The checklist javascript code allows implementing a performance
analysis tree in a manner similar to toplev
(https://github.com/andikleen/pmu-tools/wiki/toplev-manual) to
investigate performance issues.  The top of the tree is most general
overview and the lower levels are more detail specific analysis.  Only
certain parts of the tree are investigated in more detail based on the
the analysis results of the upper levels.

Each node in the tree has a plain text description of the performance
issue and a function to score the likelihood that this particular
issue exists.  Valid functions values range from 0 (definitely doesn't
exist) to 1 (definitely exists).

There is currently a simple tree in the patch that splits performance
issues to cpu, memory, networking, and storage.  The patch also
includes code to print out the tree and do a search of the tree based
on the values the node functions evaluate to.  However, the likelihood
values are not actual performance metrics at the moment.