qa/089: improve test stability
After recent "log early" changes to pmlogger, the archive used in
this test (qa/src/bar) that is created on the fly as needed may
contain some variablility in the timestamp intervals at the
beginning of the archive.  Skip this first few samples to get
into the stable part of the archive.