pcp-atop: fix -b/-e option handling to call libpcp interfaces correctly
Marko found the pcp-atop -b/-e options were not working as documented.
Turned out to be a mismatch between PMAPI wrapping and the conventions
for the called time routine.  Fix for both atop and atopsar, and add a
regression test.

Also update the quick ref guide to mention pcp atopsar, and remove the
reference to requiring a large window for atop - one of the advantages
of using the upstream atop code was those bugs got fixed.