Compute rates for individual instance and min/max values for all instances
PCP returns monotonically increasing counter for a number of metrics
and one would prefer to have rates to better gauge what is happening
over the interval.  A function has been written to compute the rate
for a particular metric instance.  The are also additional functions
to provide the maximum and minimum rate for all the instances of the
metrics.

The cpu limited predicate has been updated to make use of the rate
compuations and provide some estimate if the system coule be cpu
limited.