qa: changes to the statfs/statvfs test program on Solaris
Switch from configure.ac checking for sys/vfs.h to sys/statfs.h
which is present on both Linux and Solaris (on Linux the former
just includes the latter).  Some struct statfs fields are also
missing on Solaris, and the syscall interface differs slightly,
so tweak the code to cater for that.