rpm spec file: changes for optionally built/packaged pmdas
Track the associated configure and makefile changes, so that now if
a (pmda) package Requires: some package at run-time and that package
is not installed at build time, then that (pmda) package is not
created.

For all distros this is a no-op change, because their build environments
typically have all packages installed.

It makes a difference for older build environments where we may not
have packages available, and so building a PCP package that cannot be
installed is not helful (and breaks automated QA).