debian build: refine Makepkgs-configure integration
Motivated by the --without-optimization changes that were NQR.

It turns out that any tweaking of the options to configure in
Makepkgs is used the first time configure is run from Makepkgs,
but was ignored in the deeper use of configure from within the
Debian build and package script.

Clean this up so both invocations of configure use the same
options.