qa/975: change "match" filter
Instead of an absolute +/- 0.000002 (that sometimes failed due to
arithmetic errors when the values are large), change to +/- 2% if
both values are not zero, else < 0.002 if one of the values is
zero.

In all the previous real code failures, the values have been off by
a mile (like x2 difference), so this should be fine.