qa: reduce differences between qa and testsuite - part 1
This is a series of commits aimed at making the contents of the
testsuite directory as close as possible to the contents of the qa
directory after a "make clean".

The objective is to maximize the chances that running QA will produce
the same results from in either the (installed) testsuite directory
or the (git) qa directory.