qa/251: move from dynamic to static archives
This test is the handling of missing volumes of an archive, and will
be more stable if a static archive is used.  The recent "log early"
changes to pmlogger has meant that the dynamic archives contain a
little more non-determinism and this was triggering false failures
of the QA test.