qa: deprecate local context and unix: access via pmwebd
Local context styles automatically authenticate with an implicit
requirement that they are only ever run by local users who have
already been authenticted in some way (local login).  pmwebd was
found to be (accidentally?) circumventing that and inadvertently
exposing sensitive information to unauthenticated, remote webapi
clients.

This includes things like process command lines and memory maps -
historically the subject of CVE-2012-3419 for pmcd/pmdaproc.

Audit suggests none of the main webapps actually use these modes
of access, by default, so this is not expected to be a big issue
for anyone.  However, just in case, a back-compat preserving "-P"
option ("permissive" mode) is provided for pmwebd, allowing one
to opt-in to the suspect behavior if confident that it is safe to
do so in a specific environment.  Default is secure mode though.

Test qa/1042 and test_webprocesses.python exercise both modes.