src/pmlogger/rc_pmlogger: rework "chkconfig" logic
If "chkconfig" (or the moral equivalent thereof, as hidden away
in is_chkconfig_on() within rc-proc.sh) is "off", then do not
start pmlogger.  Previous behaviour was to warn, but start
anyway.

And when stopping, if "chkconfig" is "off" don't warn if a
pmlogger process cannot be found, as that is now expected.