build: remove all use of dynamic library loading during build
Rework RUN_IN_BUILD_ENV so that it no longer uses dynamic loader
mechanisms to insert freshly built PCP libraries in front of the
freshly build PCP utilities it uses.  Instead we build variants
(named foo.static) of the few tools needed during the build with
static PCP libraries, which we happened to build already.

This is all brought on by Mac OS X (El Capitan) security changes
which remove DYLD_* variables in SIP mode (the default), and in
particular: https://forums.developer.apple.com/thread/9233

This is the first of a series of fixes for:
https://github.com/performancecopilot/pcp/issues/48