src/include/GNUmakefile: NetBSD changes
Due to the NetBSD packaging rules, PCP bits will generally not be
installed in /var/lib/pcp and /etc/pcp.

Mostly we handle this OK, but we have to bend the rules for a couple
of files ... pcp.conf and pcp.env need to be in /etc or we are up
for a whole truckload of pain.  So force these ones into /etc, not
$(PCP_ETC_DIR), for NetBSD (where PCP_ETC_DIR is /usr/pkg/etc/pcp).