pmfg archive mode improvement: seek to end for lookup retry
The "pmclient -a" test qa/079 highlighted a problematic case both for
normal pmapi and fetchgroups: namely the lookup for metrics/indoms may
not be valid at the time origin when the operations are attempted.
The fetchgroup mechanism is now extended to retry such lookups by
temporarily seeking to the end of archive.  This can still fail, if
the indom instance existed only briefly in the middle of the archive
timeline.  The user is made aware of this possibility by new blurbage
in the man page.