build: refactor pmda.h to expose pmdaroot internals separately
Some code refactoring to prepare the way for the next round of
pmdaroot changes (starting PMDAs on behalf of PMCD).  Its become
clear that increasingly large amounts of detailed internal info
(root PDU macros, structures, functions) that do not need to be
visible to most PMDAs should be handled via a separate pmdaroot
header.  This is arguably an API/ABI break, but being realistic
noone is ever going to be affected (and pmdaroot.h is available
still anyway).