pmrep: adjust examples for RHBZ#1289912
The easiest solution for pmrep command line handling needs a change how
the context is being setup. But looks like it also revealed a corner
case issue in Python PMAPI command like option handling. Adjust the
examples to make to err on the safe side for the time being regardless
of the actual root cause. Adjust the allcache derived examples while at
it.