pmrep: further command line options guards, qa fix
Noticed qa/1069 started failing with commit 876645e038 as it
allows a None value to be passed into the context builder.

@@ -1,9 +1,13 @@
 QA output created by 1069
 == basic live mode reporting
-          s.l.one
-
-HH:MM:SS      N/A
-HH:MM:SS        1
+Traceback (most recent call last):
+  File "/bin/pmrep", line 1320, in <module>
+    P.connect()
+  File "/bin/pmrep", line 1313, in connect
+    self.context = pmapi.pmContext.fromOptions(self.opts, sys.argv, self.opts.pmGetOptionContext())
+  File "/usr/lib64/python3.3/site-packages/pcp/pmapi.py", line 1086, in fromOptions
+    if (typed <= 0):
+TypeError: unorderable types: NoneType() <= int()
 == basic archive mode reporting
           s.seconds  s.milliseconds
                util            util

Guard against this by making the pmapi wrapper code allow None,
and also guard against the BZ 1289912 situation by enforcing the
use of POSIX style argument parsing (this will need a more complex
fix in the python wrapper - leave that for next time, at this late
stage of a release).