pmrep: fix command line parsing, again
ef4312 tried to address the issue of not being able to provide options
after metrics but it turned out that that route hit the wall elsewhere
(RHBZ#1289912) and it also caused a horrible regression where command
line options did not override configuration file parameters any more.

Revert back to what we had earlier in use for a long time and sort
out things in a more robust fashion after first release as needed.