src/libpcp/context.c: tweak pmNewContext() when re-using an existing pmcd connection
There is fast path (and fd saving) optimization in pmNewContext()
where we don't establish a new connection to pmcd when we're asked
to create a subsequent PMAPI context to the same pmcd host and port
used to create an existing PMAPI context.  Multiple PMAPI contexts
are then multiplexed over the same socket connection.  This all
works as designed.

But as shown in https://github.com/performancecopilot/pcp/issues/14
this produces slightly odd behaviour when pmcd has died between the
initial pmNewContext() call and the later one, but the client has not
sent a PDU and so libpcp does not update the state of the connection
to pmcd and will try to re-use the socket that is not going to work.

Since we cannot rely on any socket-level service to let us know
the remote end of a socket has been closed, we need a "ping" ... we
don't have a ping PDU per se (and adding one would introduce version
compatibility issues) so we use a (bad) pmDesc request PDU with an
expected PM_ERR_PMID error PDU response ... if this happens, pmcd is
well, otherwise go down the slow path and try to build a new socket
connection to (presumably a new) pmcd.